RLHF_pipeline_automisation/functions/requirements.txt
firebase-admin>=6.0.0
firebase-functions>=0.2.0,<0.5.0   
google-cloud-pubsub>=2.13.0
google-cloud-aiplatform>=1.25.0
